+++ src/controllers/User.js
@@ -2,18 +2,19 @@ const { flatten } = require("mongo-dot-notation");
 const { sign } = require("jsonwebtoken");
 const passport = require("passport");
 const bcrypt = require("bcrypt");
-const User = require("../models/User");
-const { generateError, sendError } = require("../helpers/Error");
-const { checkRouteParameters } = require("../helpers/Express");
+const User = require("../db/models/User");
+const { generateError, sendError } = require("../helpers/functions/Error");
+const { checkRequestData } = require("../helpers/functions/Express");
 const Config = require("../../config");
 
 exports.create = async(data, options = {}) => {
-    const { lean } = options;
+    const { toJSON } = options;
+    delete options.toJSON;
     if (!Array.isArray(data)) {
         options = null;
     }
     const user = await User.create(data, options);
-    return lean ? user.toObject() : user;
+    return toJSON ? user.toObject() : user;
 };
 
 exports.find = async(search, projection, options = {}) => await User.find(search, projection, options);
@@ -21,12 +22,15 @@ exports.find = async(search, projection, options = {}) => await User.find(search
 exports.findOne = async(search, projection, options = {}) => await User.findOne(search, projection, options);
 
 exports.findOneAndUpdate = async(search, data, options = {}) => {
+    const { toJSON } = options;
+    delete options.toJSON;
     options.new = options.new === undefined ? true : options.new;
     const user = await this.findOne(search);
     if (!user) {
         throw generateError("NOT_FOUND", `User not found`);
     }
-    return await User.findOneAndUpdate(search, flatten(data), options);
+    const updatedUser = await User.findOneAndUpdate(search, flatten(data), options);
+    return toJSON ? updatedUser.toJSON() : updatedUser;
 };
 
 exports.generateSaltAndHash = body => new Promise((resolve, reject) => {
@@ -48,9 +52,9 @@ exports.generateSaltAndHash = body => new Promise((resolve, reject) => {
 
 exports.postUser = async(req, res) => {
     try {
-        const { body } = checkRouteParameters(req);
+        const { body } = checkRequestData(req);
         await this.generateSaltAndHash(body);
-        const newUser = await this.create(body, { lean: true });
+        const newUser = await this.create(body, { toJSON: true });
         delete newUser.password;
         res.status(200).json(newUser);
     } catch (e) {
@@ -67,9 +71,19 @@ exports.getUsers = async(req, res) => {
     }
 };
 
+exports.getUser = async(req, res) => {
+    try {
+        const { params } = checkRequestData(req);
+        const users = await this.findOne({ _id: params.id }, "-password");
+        res.status(200).json(users);
+    } catch (e) {
+        sendError(res, e);
+    }
+};
+
 exports.login = (req, res) => {
     try {
-        checkRouteParameters(req);
+        checkRequestData(req);
         passport.authenticate("local", { session: false }, (err, user, info) => {
             if (err || !user) {
                 return res.status(401).json(generateError("BAD_CREDENTIALS", info));