+++ src/db/models/Game.js
@@ -1,0 +1,4 @@
+const mongoose = require("mongoose");
+const gameSchema = require("../schemas/Game");
+const gameModel = mongoose.model("games", gameSchema);
+module.exports = gameModel;