+++ src/helpers/functions/Error.js
@@ -1,15 +1,7 @@
-const Error = require("../classes/Error");
-const Config = require("../../config");
+const Error = require("../../classes/Error");
+const Config = require("../../../config");
 
-exports.generateError = (errorType, error) => {
-    if (typeof error === "string") {
-        error = [{ msg: error }];
-    }
-    if (!Array.isArray(error)) {
-        error = [error];
-    }
-    return new Error(errorType, error);
-};
+exports.generateError = (errorType, data) => new Error(errorType, data);
 
 exports.sendUniqueViolationError = (res, e) => {
     const myRegexp = new RegExp(`${Config.db.name}.([a-z]+)`);
@@ -24,10 +16,10 @@ exports.sendUniqueViolationError = (res, e) => {
 exports.sendError = (res, e) => {
     if (e && res.headersSent) {
         console.log(e);
-    } else if (e && e.response && e.response.data && e.response.data.statusCode) {
-        res.status(e.response.data.statusCode).json(this.generateError(e.response.data.type, e.response.data.error || e.response.data.errors));
+    } else if (e && e.response && e.response.data && e.response.data.HTTPCode) {
+        res.status(e.response.data.HTTPCode).json(this.generateError(e.response.data.type, e.response.data.error || e.response.data.errors));
     } else if (e && e instanceof Error) {
-        res.status(e.statusCode).json(e);
+        res.status(e.HTTPCode).json(e);
     } else if (e && e.code === 11000) {
         this.sendUniqueViolationError(res, e);
     } else if (e && e.constructor && e.constructor.name === "JsonWebTokenError") {