+++ src/helpers/functions/Test.js
@@ -1,0 +1,5 @@
+const mongoose = require("mongoose");
+
+exports.resetDatabase = done => {
+    Promise.all(mongoose.connection.modelNames().map(model => mongoose.model(model).deleteMany())).then(() => done());
+};