+++ tests/e2e/app.test.js
@@ -1,4 +1,4 @@
-const app = require("../app");
+const app = require("../../app");
 const chai = require("chai");
 const { describe, it } = require("mocha");
 const chaiHttp = require("chai-http");
@@ -7,7 +7,7 @@ chai.use(chaiHttp);
 const { expect } = chai;
 
 describe("Testing main route", () => {
-    it("Welcomes user with API name", done => {
+    it("👋 Welcomes user with API name", done => {
         chai.request(app)
             .get("/")
             .end((err, res) => {