+++ tests/e2e/index.js
@@ -1,0 +1,3 @@
+require("./app.test");
+require("./user");
+require("./game");